agent/pyproject.toml
[project]
name = "deepagent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-langgraph>=0.0.23",
    "asyncpg>=0.31.0",
    "copilotkit>=0.1.76",
    "fastapi>=0.115.14",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
